Added Route Group

+++ app/Http/Controllers/IdeaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Idea;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class IdeaController extends Controller
 {
@@ -20,6 +21,7 @@ class IdeaController extends Controller
         ]);
 
         Idea::create([
+            "user_id" => Auth::user()->id,
             "content" => request()->get("content", ""),
         ]);
 
@@ -28,6 +30,10 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if(Auth::user()->id !== $idea->user_id) {
+            abort(404);
+        }
+
         $editing = true;
 
         return view("IdeaPages.Show", compact("idea", "editing"));
@@ -35,6 +41,10 @@ class IdeaController extends Controller
 
     public function update(Idea $idea)
     {
+        if(Auth::user()->id !== $idea->user_id) {
+            abort(404);
+        }
+
         request()->validate([
             "content"=> "required|min:5|max:240",
         ]);
@@ -53,6 +63,10 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if(Auth::user()->id !== $idea->user_id) {
+            abort(404);
+        }
+
         $idea->delete();
         return redirect()->route("idea.dashboard")->with("success","Idea was deleted successfully.");
     }
